Correção das Entidades e Relacionamentos

+++ src/simulado.shared/Entidades/SessaoSimulado.cs
@@ -6,7 +6,7 @@ public class SessaoSimulado :  Entidade
     public Guid UsuarioId { get; set; }
     public DateTime DataInicio { get; set; }
     public DateTime? DataFim { get; set; }
-    public int Pontuacao { get; set; }
+    public double Pontuacao { get; set; }
     public Simulado Simulado { get; set; }
     public IEnumerable<RespostaUsuario> RespostasUsuario { get; set; }
 }
